requirements.txt
Django==4.2
djangorestframework==3.14.0
elasticsearch==8.6.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-environ==0.11.2
gunicorn==21.2.0
uvicorn==0.27.0.post1
fastapi==0.110.0
pillow==10.2.0
python-dotenv==1.0.1
